+++ src/Compilers/Core/Portable/PEWriter/MetadataWriter.PortablePdb.cs
@@ -709,103 +709,5 @@ namespace Microsoft.Cci
         }
 
         #endregion
-
-        #region Table Serialization
-
-        private void SerializeDocumentTable(BlobBuilder writer, MetadataSizes metadataSizes)
-        {
-            foreach (var row in _documentTable)
-            {
-                writer.WriteReference((uint)_debugHeapsOpt.ResolveBlobIndex(row.Name), metadataSizes.BlobIndexSize);
-                writer.WriteReference(row.HashAlgorithm, metadataSizes.GuidIndexSize);
-                writer.WriteReference((uint)_debugHeapsOpt.ResolveBlobIndex(row.Hash), metadataSizes.BlobIndexSize);
-                writer.WriteReference(row.Language, metadataSizes.GuidIndexSize);
-            }
-        }
-
-        private void SerializeMethodDebugInformationTable(BlobBuilder writer, MetadataSizes metadataSizes)
-        {
-            foreach (var row in _methodDebugInformationTable)
-            {
-                writer.WriteReference(row.Document, metadataSizes.DocumentIndexSize);
-                writer.WriteReference((uint)_debugHeapsOpt.ResolveBlobIndex(row.SequencePoints), metadataSizes.BlobIndexSize);
-            }
-        }
-
-        private void SerializeLocalScopeTable(BlobBuilder writer, MetadataSizes metadataSizes)
-        {
-            foreach (var row in _localScopeTable)
-            {
-                writer.WriteReference(row.Method, metadataSizes.MethodDefIndexSize);
-                writer.WriteReference(row.ImportScope, metadataSizes.ImportScopeIndexSize);
-                writer.WriteReference(row.VariableList, metadataSizes.LocalVariableIndexSize);
-                writer.WriteReference(row.ConstantList, metadataSizes.LocalConstantIndexSize);
-                writer.WriteUInt32(row.StartOffset);
-                writer.WriteUInt32(row.Length);
-            }
-        }
-
-        private void SerializeLocalVariableTable(BlobBuilder writer, MetadataSizes metadataSizes)
-        {
-            foreach (var row in _localVariableTable)
-            {
-                writer.WriteUInt16(row.Attributes);
-                writer.WriteUInt16(row.Index);
-                writer.WriteReference((uint)_debugHeapsOpt.ResolveStringIndex(row.Name), metadataSizes.StringIndexSize);
-            }
-        }
-
-        private void SerializeLocalConstantTable(BlobBuilder writer, MetadataSizes metadataSizes)
-        {
-            foreach (var row in _localConstantTable)
-            {
-                writer.WriteReference((uint)_debugHeapsOpt.ResolveStringIndex(row.Name), metadataSizes.StringIndexSize);
-                writer.WriteReference((uint)_debugHeapsOpt.ResolveBlobIndex(row.Signature), metadataSizes.BlobIndexSize);
-            }
-        }
-
-        private void SerializeImportScopeTable(BlobBuilder writer, MetadataSizes metadataSizes)
-        {
-            foreach (var row in _importScopeTable)
-            {
-                writer.WriteReference(row.Parent, metadataSizes.ImportScopeIndexSize);
-                writer.WriteReference((uint)_debugHeapsOpt.ResolveBlobIndex(row.Imports), metadataSizes.BlobIndexSize);
-            }
-        }
-
-        private void SerializeStateMachineMethodTable(BlobBuilder writer, MetadataSizes metadataSizes)
-        {
-            foreach (var row in _stateMachineMethodTable)
-            {
-                writer.WriteReference(row.MoveNextMethod, metadataSizes.MethodDefIndexSize);
-                writer.WriteReference(row.KickoffMethod, metadataSizes.MethodDefIndexSize);
-            }
-        }
-
-        private void SerializeCustomDebugInformationTable(BlobBuilder writer, MetadataSizes metadataSizes)
-        {
-            // sort by Parent, Kind
-            _customDebugInformationTable.Sort(CustomDebugInformationRowComparer.Instance);
-
-            foreach (var row in _customDebugInformationTable)
-            {
-                writer.WriteReference(row.Parent, metadataSizes.HasCustomDebugInformationSize);
-                writer.WriteReference(row.Kind, metadataSizes.GuidIndexSize);
-                writer.WriteReference((uint)_debugHeapsOpt.ResolveBlobIndex(row.Value), metadataSizes.BlobIndexSize);
-            }
-        }
-
-        private class CustomDebugInformationRowComparer : Comparer<CustomDebugInformationRow>
-        {
-            public static readonly CustomDebugInformationRowComparer Instance = new CustomDebugInformationRowComparer();
-
-            public override int Compare(CustomDebugInformationRow x, CustomDebugInformationRow y)
-            {
-                int result = (int)x.Parent - (int)y.Parent;
-                return (result != 0) ? result : (int)x.Kind - (int)y.Kind;
-            }
-        }
-
-        #endregion
     }
 }
